pyproject.toml
[project]
name = "dealsignals"
version = "0.1.0"
description = "AI-human symbiosis research for deal analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Leonard", email = "leonard@yri.ai" }
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.40",
    "openai>=1.50",
]
parsing = [
    "pypdf>=4.0",
    "pdfplumber>=0.11",
]
rag = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
]
all = [
    "dealsignals[llm,parsing,rag]",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
ds = "dealsignals.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dealsignals"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
